requirements.txt
PyYAML>=6.0.1
openai>=1.40.0
transformers>=4.44.0
torch>=2.2.0
numpy>=1.26.0
scikit-learn>=1.5.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
gradio>=4.44.0
